LayoutTests/fast/canvas/canvas-2d-imageData-create-nonfinite-expected.txt

Test the argument bounds of canvas createImageData.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.createImageData(Infinity, Infinity) threw exception TypeError: Value Infinity is outside the range [-2147483648, 2147483647].
PASS ctx.createImageData(Infinity, 10) threw exception TypeError: Value Infinity is outside the range [-2147483648, 2147483647].
PASS ctx.createImageData(-Infinity, 10) threw exception TypeError: Value -Infinity is outside the range [-2147483648, 2147483647].
PASS ctx.createImageData(10, Infinity) threw exception TypeError: Value Infinity is outside the range [-2147483648, 2147483647].
PASS ctx.createImageData(10, -Infinity) threw exception TypeError: Value -Infinity is outside the range [-2147483648, 2147483647].
PASS ctx.createImageData(NaN, 10) threw exception TypeError: Value NaN is outside the range [-2147483648, 2147483647].
PASS ctx.createImageData(10, NaN) threw exception TypeError: Value NaN is outside the range [-2147483648, 2147483647].
PASS successfullyParsed is true

TEST COMPLETE
